--- src/utils/timestampToAgo.ts
@@ -1,6 +1,0 @@
-import { formatDistanceToNow, fromUnixTime } from 'date-fns';
-
-export const timestampToAgo = (timestamp: number) => {
-  const date = fromUnixTime(timestamp);
-  return formatDistanceToNow(date, { addSuffix: true }).toString();
-};
