--- src/utils/timestampToDate.ts
@@ -1,6 +1,0 @@
-import { fromUnixTime, lightFormat } from 'date-fns';
-
-export const timestampToDate = (timestamp: number) => {
-  const date = fromUnixTime(timestamp);
-  return lightFormat(date, 'HH:mm dd.MM.yy').toString();
-};
